doi:10.3390/ma13102373
Recycling is an important habit to avoid waste. This paper evaluates the performance of masonry mortar, elaborated by replacing natural sand with recycled fine aggregate (RFA) obtained from mortar. Five families of mixtures were prepared with dierent replacement proportions: 20%, 40%, 60%, and 100%. A 1:4 volumetric cement-to-aggregate ratio was used for all mixtures by experimentally adjusting the amount of water to achieve the same consistency of 175 mm. The efects of the following procedures were analyzed: (1) the use of a deconstruction technique to collect the RFA, (2) pre-wetting of the aggregates, and (3) the use of a commercial plasticizer. Experimental results show that it is possible to use this type of recycled fine aggregate as a substitute for natural sand by up to 60% in the manufacture of masonry mortar without significantly afecting
its properties.